[Feat] : ArticleSearchService 커맨드, 직군, 키워드 분리 (#95)
